v3: address testing comments
